Clean/legacy graphics api cleanup (#43)
* clean and update README

* remove blazor components

* more clean code
