+++ Drilla.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DualDrill.Server", "DualDrill.Server\DualDrill.Server.csproj", "{158B5D6E-4080-439F-AFFA-2859AC625BEA}"
 EndProject
-Project("{54A90642-561A-4BB1-A94E-469ADEE60C69}") = "DualDrill.JS", "DualDrill.JS\DualDrill.JS.esproj", "{96A0AA95-6474-41DC-ACB9-9C2F3D34097E}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DualDrill.Engine", "DualDrill.Engine\DualDrill.Engine.csproj", "{CD5B7346-7767-48CE-91C4-53FB2C063D00}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DualDrill.Common", "DualDrill.Common\DualDrill.Common.csproj", "{8561C01E-FC29-4DE6-BA11-2552D496ECC0}"
@@ -39,18 +37,6 @@ Global
 		{158B5D6E-4080-439F-AFFA-2859AC625BEA}.Release|Any CPU.Build.0 = Release|Any CPU
 		{158B5D6E-4080-439F-AFFA-2859AC625BEA}.Release|x64.ActiveCfg = Release|Any CPU
 		{158B5D6E-4080-439F-AFFA-2859AC625BEA}.Release|x64.Build.0 = Release|Any CPU
-		{96A0AA95-6474-41DC-ACB9-9C2F3D34097E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{96A0AA95-6474-41DC-ACB9-9C2F3D34097E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{96A0AA95-6474-41DC-ACB9-9C2F3D34097E}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
-		{96A0AA95-6474-41DC-ACB9-9C2F3D34097E}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{96A0AA95-6474-41DC-ACB9-9C2F3D34097E}.Debug|x64.Build.0 = Debug|Any CPU
-		{96A0AA95-6474-41DC-ACB9-9C2F3D34097E}.Debug|x64.Deploy.0 = Debug|Any CPU
-		{96A0AA95-6474-41DC-ACB9-9C2F3D34097E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{96A0AA95-6474-41DC-ACB9-9C2F3D34097E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{96A0AA95-6474-41DC-ACB9-9C2F3D34097E}.Release|Any CPU.Deploy.0 = Release|Any CPU
-		{96A0AA95-6474-41DC-ACB9-9C2F3D34097E}.Release|x64.ActiveCfg = Release|Any CPU
-		{96A0AA95-6474-41DC-ACB9-9C2F3D34097E}.Release|x64.Build.0 = Release|Any CPU
-		{96A0AA95-6474-41DC-ACB9-9C2F3D34097E}.Release|x64.Deploy.0 = Release|Any CPU
 		{CD5B7346-7767-48CE-91C4-53FB2C063D00}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{CD5B7346-7767-48CE-91C4-53FB2C063D00}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CD5B7346-7767-48CE-91C4-53FB2C063D00}.Debug|x64.ActiveCfg = Debug|Any CPU
