+++ DualDrill.Engine/Renderer/VolumeRenderer.cs
@@ -1,13 +1,5 @@
 ﻿using DualDrill.Engine.Services;
 using DualDrill.Graphics;
-using DualDrill.Graphics.Backend;
-using DualDrill.Graphics.Interop;
-using Silk.NET.Vulkan;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace DualDrill.Engine.Renderer;
