+++ DualDrill.Graphics/Backend/WebGPUNet/WebGPUNETBackend.cs
@@ -128,11 +128,6 @@ public sealed partial class WebGPUNETBackend : IBackend<Backend>
         throw new NotImplementedException();
     }
 
-    ValueTask<GPUDevice> IBackend<Backend>.RequestDeviceAsyncLegacy(GPUAdapter<Backend> adapter, GPUDeviceDescriptor descriptor, CancellationToken cancellation)
-    {
-        throw new NotImplementedException();
-    }
-
     public void Dispose()
     {
     }
