+++ DualDrill.Graphics/GPUAdapter.cs
@@ -3,15 +3,12 @@
 public partial interface IGPUAdapter : IDisposable
 {
     public ValueTask<GPUAdapterInfo> RequestAdapterInfoAsync(CancellationToken cancellation);
+    public ValueTask<IGPUDevice> RequestDeviceAsync(GPUDeviceDescriptor descriptor, CancellationToken cancellation);
 }
 
-public sealed partial record class GPUAdapter<TBackend>
-{
-    public ValueTask<GPUDevice> RequestDeviceAsyncLegacy(GPUDeviceDescriptor descriptor, CancellationToken cancellation)
-       => TBackend.Instance.RequestDeviceAsyncLegacy(this, descriptor, cancellation);
-}
-
-public sealed partial record class GPUAdapter<TBackend>
+public sealed partial record class GPUAdapter<TBackend>(GPUHandle<TBackend, GPUAdapter<TBackend>> Handle)
+    : IDisposable, IGPUAdapter
+    where TBackend : IBackend<TBackend>
 {
     public async ValueTask<IGPUDevice> RequestDeviceAsync(GPUDeviceDescriptor descriptor, CancellationToken cancellation)
     {
@@ -22,5 +19,11 @@ public sealed partial record class GPUAdapter<TBackend>
     {
         return TBackend.Instance.RequestAdapterInfoAsync(this, cancellation);
     }
+
+    public void Dispose()
+    {
+        TBackend.Instance.DisposeHandle(Handle);
+    }
 }
+
