+++ DualDrill.Graphics/GPUDevice.cs
@@ -1,6 +1,4 @@
-﻿using DualDrill.Graphics.Interop;
-
-namespace DualDrill.Graphics;
+﻿namespace DualDrill.Graphics;
 
 public partial interface IGPUDevice : IDisposable
 {
@@ -140,117 +138,3 @@ public sealed partial record class GPUDevice<TBackend>(GPUHandle<TBackend, GPUDe
 }
 
 
-
-
-public sealed partial class GPUDevice
-{
-    //public GPURenderPipeline CreateRenderPipeline(GPURenderPipelineDescriptor descriptor)
-    //{
-    //    return GPURenderPipeline.Create(this, descriptor);
-    //}
-
-    public unsafe GPUCommandEncoder CreateCommandEncoder(GPUCommandEncoderDescriptor descriptor)
-    {
-        if (descriptor.Label is not null)
-        {
-            throw new NotImplementedException();
-        }
-        WGPUCommandEncoderDescriptor nativeDescriptor = new()
-        {
-        };
-        return new(WGPU.DeviceCreateCommandEncoder(Handle, &nativeDescriptor));
-    }
-
-    //public unsafe GPUSampler CreateSampler(GPUSamplerDescriptor descriptor)
-    //{
-    //    WGPUSamplerDescriptor nativeDescriptor = new()
-    //    {
-    //        addressModeU = descriptor.AddressModeU,
-    //        addressModeV = descriptor.AddressModeV,
-    //        addressModeW = descriptor.AddressModeW,
-    //        magFilter = descriptor.MagFilter,
-    //        minFilter = descriptor.MinFilter,
-    //        mipmapFilter = descriptor.MipmapFilter,
-    //        lodMinClamp = descriptor.LodMinClamp,
-    //        lodMaxClamp = descriptor.LodMaxClamp,
-    //        compare = descriptor.Compare,
-    //        maxAnisotropy = descriptor.MaxAnisotropy
-    //    };
-    //    return new(WGPU.DeviceCreateSampler(Handle, &nativeDescriptor));
-    //}
-
-    public unsafe void Poll()
-    {
-        _ = WGPU.DevicePoll(Handle, 0, null);
-    }
-
-    internal async Task PollAndWaitOnTaskAsync(Task target)
-    {
-        unsafe void DoPoll()
-        {
-            _ = WGPU.DevicePoll(Handle, 0, null);
-        }
-        while (!target.IsCompleted)
-        {
-            DoPoll();
-            await Task.Yield();
-        }
-    }
-
-
-    public unsafe GPUQueue GetQueue()
-    {
-        return new(WGPU.DeviceGetQueue(Handle));
-    }
-
-    //public unsafe GPUTexture CreateTexture(GPUTextureDescriptor descriptor)
-    //{
-
-    //    using var label = NativeStringRef.Create(descriptor.Label);
-    //    var viewFormats = stackalloc GPUTextureFormat[descriptor.ViewFormats.Length];
-    //    for (var i = 0; i < descriptor.ViewFormats.Length; i++)
-    //    {
-    //        viewFormats[i] = descriptor.ViewFormats.Span[i];
-    //    }
-    //    WGPUTextureDescriptor native = new()
-    //    {
-    //        label = (sbyte*)label.Handle,
-    //        usage = (uint)descriptor.Usage,
-    //        dimension = descriptor.Dimension,
-    //        size = descriptor.Size,
-    //        format = descriptor.Format,
-    //        mipLevelCount = (uint)descriptor.MipLevelCount,
-    //        sampleCount = (uint)descriptor.SampleCount,
-    //        viewFormatCount = (nuint)descriptor.ViewFormats.Length,
-    //        viewFormats = descriptor.ViewFormats.Length > 0 ? viewFormats : null,
-    //    };
-    //    return new GPUTexture(WGPU.DeviceCreateTexture(Handle, &native));
-    //}
-
-    //public unsafe GPUBuffer CreateBuffer(GPUBufferDescriptor descriptor)
-    //{
-    //    var alignedSize = (descriptor.Size + 3UL) & ~3UL;
-    //    //Debug.Assert(descriptor.Size == alignedSize, "Buffer byte size should be multiple of 4");
-    //    WGPUBufferDescriptor nativeDescriptor = new()
-    //    {
-    //        mappedAtCreation = (GPUBool)descriptor.MappedAtCreation,
-    //        size = alignedSize,
-    //        usage = (uint)descriptor.Usage,
-    //    };
-    //    var handle = WGPU.DeviceCreateBuffer(Handle, &nativeDescriptor);
-    //    return new(handle);
-    //}
-
-    //public unsafe IGPUBindGroup CreateBindGroup(GPUBindGroupDescriptor descriptor)
-    //{
-      
-    //}
-
-    //public unsafe GPUBindGroupLayout CreateBindGroupLayout(GPUBindGroupLayoutDescriptor descriptor)
-    //{
-    //      }
-
-    //public unsafe GPUPipelineLayout CreatePipelineLayout(GPUPipelineLayoutDescriptor descriptor)
-    //{
-    //       }
-}
