+++ DualDrill.Graphics/GPUQueue.cs
@@ -1,6 +1,4 @@
-﻿using DualDrill.Graphics.Interop;
-using Silk.NET.Vulkan;
-using System.Runtime.CompilerServices;
+﻿using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 namespace DualDrill.Graphics;
@@ -56,108 +54,3 @@ where TBackend : IBackend<TBackend>
     }
 }
 
-
-public sealed partial class GPUQueue
-{
-    public unsafe void Submit(ReadOnlySpan<GPUCommandBuffer> buffers)
-    {
-        var native = stackalloc WGPUCommandBufferImpl*[buffers.Length];
-        for (var i = 0; i < buffers.Length; i++)
-        {
-            native[i] = buffers[i].Handle;
-        }
-        WGPU.QueueSubmit(Handle, (uint)buffers.Length, native);
-    }
-
-    public unsafe void WriteBuffer<T>(GPUBuffer buffer, ulong bufferOffset, ReadOnlySpan<T> data)
-        where T : unmanaged
-    {
-        var byteSpan = MemoryMarshal.Cast<T, byte>(data);
-        if (byteSpan.Length % 4 != 0)
-        {
-            throw new InvalidOperationException($"WriteBuffer data length must be a multiple of 4 bytes, got {byteSpan.Length}");
-        }
-        var ptr = Unsafe.AsPointer(ref MemoryMarshal.GetReference(byteSpan));
-        WGPU.QueueWriteBuffer(Handle, buffer.Handle, bufferOffset, ptr, (nuint)byteSpan.Length);
-    }
-
-    //public unsafe void WriteTexture(
-    //    GPUImageCopyTexture destination,
-    //    ReadOnlySpan<byte> data,
-    //    GPUImageDataLayout layout,
-    //    GPUExtent3D extent)
-    //{
-    //    var nativeDestination = new WGPUImageCopyTexture
-    //    {
-    //        aspect = destination.Aspect,
-    //        mipLevel = (uint)destination.MipLevel,
-    //        origin = new()
-    //        {
-    //            x = (uint)destination.Origin.X,
-    //            y = (uint)destination.Origin.Y,
-    //            z = (uint)destination.Origin.Z,
-    //        },
-    //        texture = destination.Texture.Handle,
-    //    };
-    //    var nativeLayout = new WGPUTextureDataLayout
-    //    {
-    //        offset = layout.Offset,
-    //        bytesPerRow = (uint)layout.BytesPerRow,
-    //        rowsPerImage = (uint)layout.RowsPerImage,
-    //    };
-    //    var nativeExtend = new WGPUExtent3D
-    //    {
-    //        width = (uint)extent.Width,
-    //        height = (uint)extent.Height,
-    //        depthOrArrayLayers = (uint)extent.DepthOrArrayLayers
-    //    };
-    //    var ptr = Unsafe.AsPointer(ref MemoryMarshal.GetReference(data));
-    //    WGPU.QueueWriteTexture(
-    //        Handle,
-    //        &nativeDestination,
-    //        ptr,
-    //        (nuint)data.Length,
-    //        &nativeLayout,
-    //        &nativeExtend
-    //    );
-    //}
-
-    [UnmanagedCallersOnly(CallConvs = [typeof(CallConvCdecl)])]
-    unsafe static void QueueWorkDone(GPUQueueWorkDoneStatus status, void* data)
-    {
-        var handle = GCHandle.FromIntPtr((nint)data);
-        //var target = (TaskCompletionSource<WGPUQueueWorkDoneStatus>)handle.Target;
-        //target.SetResult(status);
-        var action = (Action?)handle.Target;
-        if (action is not null)
-        {
-            action();
-        }
-        handle.Free();
-    }
-
-    [UnmanagedCallersOnly(CallConvs = [typeof(CallConvCdecl)])]
-    unsafe static void QueueWorkDoneAsync(GPUQueueWorkDoneStatus status, void* data)
-    {
-        var handle = GCHandle.FromIntPtr((nint)data);
-        var target = handle.Target as TaskCompletionSource<GPUQueueWorkDoneStatus> ?? throw new GraphicsApiException($"GCHandle({(nint)data:X}) failed to recover target");
-        target.SetResult(status);
-        handle.Free();
-    }
-
-    public unsafe Task WaitSubmittedWorkDoneAsync(CancellationToken cancellation = default)
-    {
-        var tcs = new TaskCompletionSource<GPUQueueWorkDoneStatus>(cancellation);
-        var handle = GCHandle.ToIntPtr(GCHandle.Alloc(tcs));
-        WGPU.QueueOnSubmittedWorkDone(Handle, &QueueWorkDoneAsync, (void*)handle);
-        return tcs.Task;
-    }
-    public unsafe void OnSubmittedWorkDone(Action next)
-    {
-        //var tcs = new TaskCompletionSource<WGPUQueueWorkDoneStatus>();
-        //var handle = GCHandle.ToIntPtr(GCHandle.Alloc(tcs));
-        var handle = GCHandle.ToIntPtr(GCHandle.Alloc(next));
-        WGPU.QueueOnSubmittedWorkDone(Handle, &QueueWorkDone, (void*)handle);
-        //return tcs.Task;
-    }
-}
