+++ DualDrill.Graphics/GPURequestAdapterOptions.cs
@@ -5,5 +5,5 @@ public partial struct GPURequestAdapterOptions
     public bool ForceFallbackAdapter { get; set; }
     public required GPUPowerPreference PowerPreference { get; set; }
     public GPUBackendType BackendType { get; set; }
-    public GPUSurface? CompatibleSurface { get; set; }
+    public IGPUSurface? CompatibleSurface { get; set; }
 }
