+++ DualDrill.Graphics/GPUTexture.cs
@@ -1,6 +1,4 @@
-﻿using DualDrill.Graphics.Interop;
-
-namespace DualDrill.Graphics;
+﻿namespace DualDrill.Graphics;
 
 public partial interface IGPUTexture : IDisposable
 {
@@ -50,25 +48,3 @@ public sealed partial record class GPUTexture<TBackend>(GPUHandle<TBackend, GPUT
 }
 
 
-
-public sealed partial class GPUTexture
-{
-    public unsafe GPUTextureView CreateView(GPUTextureViewDescriptor? descriptor = null)
-    {
-        if (!descriptor.HasValue)
-        {
-            return new(WGPU.TextureCreateView(Handle, null));
-        }
-        WGPUTextureViewDescriptor native = new()
-        {
-            format = descriptor.Value.Format,
-            dimension = descriptor.Value.Dimension,
-            baseMipLevel = (uint)descriptor.Value.BaseMipLevel,
-            mipLevelCount = (uint)descriptor.Value.MipLevelCount,
-            baseArrayLayer = (uint)descriptor.Value.BaseArrayLayer,
-            arrayLayerCount = (uint)descriptor.Value.ArrayLayerCount,
-            aspect = descriptor.Value.Aspect
-        };
-        return new(WGPU.TextureCreateView(Handle, &native));
-    }
-}
