+++ DualDrill.Graphics/GPUTextureView.cs
@@ -1,6 +1,4 @@
-﻿using DualDrill.Graphics.Interop;
-
-namespace DualDrill.Graphics;
+﻿namespace DualDrill.Graphics;
 
 
 public partial interface IGPUTextureView : IDisposable
@@ -20,16 +18,3 @@ public sealed partial record class GPUTextureView<TBackend>(GPUHandle<TBackend, 
 
 
 
-public sealed partial class GPUTextureView : IDisposable
-{
-    internal NativeHandle<WGPUNativeApiInterop, WGPUTextureViewImpl> Handle { get; }
-    internal unsafe GPUTextureView(WGPUTextureViewImpl* handle)
-    {
-        Handle = new(handle);
-    }
-
-    public unsafe WGPUTextureViewImpl* NativePointer => Handle; public void Dispose()
-    {
-        Handle.Dispose();
-    }
-}
