+++ DualDrill.Graphics/IBackend.cs
@@ -5,11 +5,6 @@ public partial interface IBackend<TBackend>
        where TBackend : IBackend<TBackend>
 {
     public abstract static TBackend Instance { get; }
-    internal ValueTask<GPUDevice> RequestDeviceAsyncLegacy(
-        GPUAdapter<TBackend> adapter,
-        GPUDeviceDescriptor descriptor,
-        CancellationToken cancellation
-    );
     internal GPUTextureView<TBackend> CreateTextureView(GPUTexture<TBackend> texture, GPUTextureViewDescriptor descriptor);
 
     internal void Poll(GPUDevice<TBackend> device);
