--- DualDrill.Graphics/Resources.Gen.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection.Metadata;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DualDrill.Graphics;
-
-internal class Resources
-{
-}
