--- DualDrill.Server/Components/Pages/DesktopBrowserClient.razor
@@ -1,14 +1,0 @@
-﻿@using Shared;
-@rendermode @(new InteractiveServerRenderMode(prerender: false))
-
-<h3>Desktop Client @ClientId</h3>
-@if (Connecting)
-{
-    <span>Connecting ... </span>
-}
-<div>
-    <span> scale </span>
-    <input type="range" min="0.1" max="2.0" step="0.1" @bind-value="Scale" @bind-value:event="oninput" />
-    <span> @Scale </span>
-</div>
-<button @onclick="Emit">Emit</button>
