--- DualDrill.Server/Components/Pages/DualDrillBrowserClient.razor
@@ -1,5 +1,0 @@
-﻿@page "/desktop"
-<h3>Browser Client</h3>
-<DualDrill.Client.Pages.ClientMain></DualDrill.Client.Pages.ClientMain>
-<DesktopBrowserClient></DesktopBrowserClient>
-
