--- DualDrill.Server/Components/Pages/Home.razor
@@ -1,5 +1,0 @@
-﻿@page "/"
-
-<PageTitle>DualDrill - Home</PageTitle>
-
-<h1>Dual Drill Home</h1>
