--- DualDrill.Server/Components/Pages/WebViewClient.razor
@@ -1,3 +1,0 @@
-﻿@page "/webview2"
-<h3>WebView Client</h3>
-
