--- DualDrill.Server/Components/Shared/PeerClientConnection.razor
@@ -1,18 +1,0 @@
-﻿<h5>PeerClient @(PeerClient?.Uri)</h5>
-<button @onclick="StateHasChanged">Force Update</button>
-@if (!IsConnected)
-{
-    <button @onclick="Connect">Connect</button>
-}
-else
-{
-    <span>Connected</span>
-}
-@if (SelfMediaStream is not null && BrowserRTCPeerConnectionPair is not null)
-{
-    <button disabled="@Sending" @onclick="SendStream">Send</button>
-}
-@if (PeerMediaStream is not null)
-{
-    <button @onclick="SelectAsShownPeerStream">Select Peer Video</button>
-}
