--- DualDrill.Server/Components/Shared/VideoViewer.razor
@@ -1,23 +1,0 @@
-﻿@using DualDrill.Engine.BrowserProxy;
-
-<video id="selfVideo" @ref="SelfVideoElement" autoplay playsinline></video>
-
-@code {
-    ElementReference SelfVideoElement;
-
-    [Parameter]
-    public JSMediaStreamProxy? MediaStream { get; set; }
-
-    [CascadingParameter]
-    public JSClientModule? ClientModule { get; set; }
-
-    protected override async Task OnParametersSetAsync()
-    {
-        await base.OnParametersSetAsync();
-        if (ClientModule is not null)
-        {
-            await using var videoElementRef = await ClientModule.CreateJSObjectReferenceAsync(SelfVideoElement).ConfigureAwait(false);
-            await ClientModule.SetVideoElementStreamAsync(videoElementRef, MediaStream?.Reference);
-        }
-    }
-}
