+++ DualDrill.Server/DualDrill.Server.csproj
@@ -14,7 +14,7 @@
 	<Target Name="PreBuild" BeforeTargets="Build">
 		<Exec Command="echo Running pnpm install" />
 		<Exec Command="pnpm install" WorkingDirectory="..\DualDrill.JS" />
-		<Exec Command ="echo running node .\esbuild.mjs" />
+		<Exec Command="echo running node .\esbuild.mjs" />
 		<Exec Command="node .\esbuild.mjs" WorkingDirectory="..\DualDrill.JS" />
 	</Target>
